+++ src/Pim/Bundle/CatalogBundle/DependencyInjection/PimCatalogExtension.php
@@ -36,9 +36,10 @@ class PimCatalogExtension extends Extension
         $loader->load('services.yml');
         $loader->load('event_listeners.yml');
         $loader->load('managers.yml');
+        $loader->load('builders.yml');
         $loader->load('attribute_types.yml');
         $loader->load('attribute_constraint_guessers.yml');
-        $loader->load('factory.yml');
+        $loader->load('factories.yml');
         $loader->load('entities.yml');
         $loader->load('repositories.yml');