+++ src/Pim/Bundle/TransformBundle/Transformer/Property/MediaTransformer.php
@@ -6,7 +6,6 @@ use Symfony\Component\HttpFoundation\File\Exception\FileNotFoundException;
 use Symfony\Component\HttpFoundation\File\File;
 use Pim\Bundle\TransformBundle\Exception\PropertyTransformerException;
 use Pim\Bundle\TransformBundle\Transformer\ColumnInfo\ColumnInfoInterface;
-use Pim\Bundle\CatalogBundle\Model\ProductMedia;
 
 /**
  * Media attribute transformer
@@ -17,6 +16,17 @@ use Pim\Bundle\CatalogBundle\Model\ProductMedia;
  */
 class MediaTransformer implements PropertyTransformerInterface, EntityUpdaterInterface
 {
+    /** @var string */
+    protected $mediaClass;
+
+    /**
+     * @param string $mediaClass
+     */
+    public function __construct($mediaClass)
+    {
+        $this->mediaClass = $mediaClass;
+    }
+
     /**
      * {@inheritdoc}
      */
@@ -48,7 +58,7 @@ class MediaTransformer implements PropertyTransformerInterface, EntityUpdaterInt
 
         $media = $object->getMedia();
         if (!$media) {
-            $media = new ProductMedia();
+            $media = new $this->mediaClass();
             $object->setMedia($media);
         }
         $media->setFile($data);